User commit

+++ src/main/java/com/example/demo/service/ValidationService.java
@@ -6,5 +6,5 @@ public interface ValidationService{
     // postdata
     ValidationEntity postData(ValidationEntity vau);
     // findbyId
-    ValidationEntity find(int id);
+    ValidationEntity find(long id);
 }
